+++ src/commands/registration.js
