+++ src/ircLineParser.js
