+++ src/numerics.js
